intersection-observer/observer-attributes.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="root"></div>

<script>
test(function() {
  var observer = new IntersectionObserver(function(e) {}, {});
  test(function() { assert_true(observer.root === null) },
       "observer.root === null");
  test(function() { assert_array_equals(observer.thresholds, [0]) },
       "observer.thresholds == [0]");
  test(function() { assert_equals(observer.rootMargin, "0px 0px 0px 0px") },
       "observer.rootMargin == '0px 0px 0px 0px'");

  var rootDiv = document.getElementById("root");
  observer = new IntersectionObserver(function(e) {}, {
    root: rootDiv,
    threshold: [0, 0.25, 0.5, 1.0],
    rootMargin: "10% 20px"
  });
  test(function() { assert_true(observer.root === rootDiv) },
       "observer.root === rootDiv");
  test(function() { assert_array_equals(observer.thresholds, [0, 0.25, 0.5, 1.0]) },
       "observer.thresholds == [0, 0.25, 0.5, 1.0]");
  test(function() { assert_equals(observer.rootMargin, "10% 20px 10% 20px") },
       "observer.rootMargin == '10% 20px 10% 20px'");
}, "Test for observer attribute getters.");

</script>
